fix(core): object status examples
