+++ apps/docs/src/app/core/component-docs/object-status/examples/object-status-default-example.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-object-status-default-example',
+    templateUrl: './object-status-default-example.component.html',
+    styleUrls: ['./object-status-examples.component.scss']
+})
+export class ObjectStatusDefaultExampleComponent {}
