+++ apps/docs/src/app/core/component-docs/object-status/examples/object-status-examples.component.ts
@@ -1,11 +1,6 @@
 import { Component } from '@angular/core';
 
-@Component({
-    selector: 'fd-object-status-default-example',
-    templateUrl: './object-status-default-example.component.html',
-    styleUrls: ['./object-status-examples.component.scss']
-})
-export class ObjectStatusDefaultExampleComponent {}
+
 
 @Component({
     selector: 'fd-object-status-text-example',
